[TM-1705] landing page (#942)
* [TM-1705] fix organizations

* [TM-1705] move to hook data

* [TM-1705] remove log
